doc,code: kg finalizes everything!
This version allows to specify test mode and size of the shared memory
as cmd parameters to automate runs in gem5. The allocator node takes
care of allocating the graph. "info:" lines are added to notify the user
what's going on in the program.

The README was updated with the correct map of the shared memory region.
The only missing feature is a cache line flush, which probably needs to
be in a different branch as it needs to have ISA specific instructions.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>